+++ packages/bluesky_cli/lib/src/command/commands.dart
@@ -13,6 +13,7 @@ import 'actor/suggestions.dart';
 import 'bsky_command.dart';
 import 'common/delete.dart';
 import 'feed/actor_feeds.dart';
+import 'feed/actor_likes.dart';
 import 'feed/create_generator.dart';
 import 'feed/custom_feed.dart';
 import 'feed/feed.dart';
@@ -78,6 +79,7 @@ final feedCommands = <BskyCommand>[
   FeedGeneratorsCommand(),
   CustomFeedCommand(),
   GeneratorInfoCommand(),
+  ActorLikesCommand(),
 ];
 
 final notificationCommands = <BskyCommand>[