+++ packages/bluesky_cli/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Note
 
+## v0.3.2
+
+- Supported `actor-likes` command. ([#700](https://github.com/myConsciousness/atproto.dart/issues/700))
+
 ## v0.3.1
 
 - Improved documents. ([#677](https://github.com/myConsciousness/atproto.dart/issues/677))